requirements_slim.txt
gensim
tqdm
pyRDF2vec
scikit-learn
umap-learn
SPARQLWrapper
jcopml
plotly
pandas
numpy
scipy
pygraphviz
matplotlib
fastapi
uvicorn[standard]
requests
python-dotenv
pymongo
ipykernel
nest_asyncio
aiohttp